requirements.txt
# Core RAG dependencies
chromadb>=0.4.22
sentence-transformers>=2.2.2
unstructured[pdf]>=0.12.0
langchain-text-splitters>=0.0.1

# LLM
google-generativeai>=0.4.0

# Web UI
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# PDF processing dependencies
pymupdf>=1.23.0
pdfplumber>=0.10.0
